refactor(runtime): move transport ownership to studio
