+++ chipcompiler/cli/command_handlers/project.py
@@ -422,7 +422,6 @@ def _run_project(
             layer_warnings.append(set_warning)
 
     # TODO: Move non-interactive project run preparation/execution into
-    # chipcompiler.runtime.project_runner.run_project or
     # chipcompiler.engine.project_run.prepare_and_run. Keep CLI ownership limited
     # to input parsing, progress renderer selection, and CommandResult mapping.
     project_state = ctx.project_state
