+++ chipcompiler/engine/subflow_events.py
@@ -1,5 +1,3 @@
-from __future__ import annotations
-
 from copy import deepcopy
 from typing import Any
