--- chipcompiler/runtime/__init__.py
@@ -1,1 +1,0 @@
-"""Private ECC runtime support for long-lived sidecar sessions."""
