--- chipcompiler/runtime/signoff_export.py
@@ -1,285 +1,0 @@
-import os
-import shutil
-import tempfile
-from pathlib import Path
-
-from chipcompiler.engine import EngineFlow, SignoffPackageOptions
-from chipcompiler.runtime.workspace_api import RuntimeApiError
-from chipcompiler.utility import json_read
-
-_REVIEW_GROUPS = (
-    ("initial", "Initial"),
-    ("config", "Config"),
-    ("harden", "Harden"),
-    ("final_design", "Final Design"),
-    ("sta", "STA"),
-    ("spef", "SPEF"),
-    ("reports", "Reports"),
-)
-
-
-def inspect_signoff_package(workspace) -> dict:
-    """Refresh current outputs, then render the home checklist contract only."""
-    EngineFlow(workspace).collect_signoff_package(
-        SignoffPackageOptions(archive=False, materialize=False, refresh_analysis=True)
-    )
-    checklist_path = Path(workspace.directory) / "home" / "checklist.json"
-    checklist_data = json_read(checklist_path)
-    if (
-        not isinstance(checklist_data, dict)
-        or checklist_data.get("schema_version") != 3
-        or checklist_data.get("kind") != "signoff_checklist"
-    ):
-        return _unavailable_review()
-
-    groups = {
-        group_id: {
-            "id": group_id,
-            "label": label,
-            "available": 0,
-            "expected": 0,
-            "blocked_details": [],
-            "attention_details": [],
-        }
-        for group_id, label in _REVIEW_GROUPS
-    }
-    for item in checklist_data.get("checklist", []):
-        if not isinstance(item, dict):
-            continue
-        group = groups[_review_group_for_item(item)]
-        group["expected"] += 1
-        if item.get("state") == "pass":
-            group["available"] += 1
-        elif item.get("blocked") is True:
-            group["blocked_details"].append(_review_detail(item))
-        else:
-            group["attention_details"].append(_review_detail(item))
-
-    review_groups = []
-    risks = []
-    for group_id, _label in _REVIEW_GROUPS:
-        group = groups[group_id]
-        blocked_details = group["blocked_details"]
-        attention_details = group["attention_details"]
-        available = group["available"]
-        expected = group["expected"]
-        if blocked_details:
-            status = "blocked"
-            summary = f"{len(blocked_details)} blocking checklist requirements"
-            risks.append(
-                {
-                    "severity": "blocked",
-                    "title": f"{group['label']} signoff requirements block export",
-                    "summary": summary,
-                    "details": blocked_details,
-                }
-            )
-            if attention_details:
-                risks.append(
-                    {
-                        "severity": "warning",
-                        "title": f"{group['label']} signoff attention",
-                        "summary": (
-                            f"{len(attention_details)} attention-only checklist requirements"
-                        ),
-                        "details": attention_details,
-                    }
-                )
-        elif attention_details:
-            status = "attention"
-            summary = f"{len(attention_details)} attention-only checklist requirements"
-            risks.append(
-                {
-                    "severity": "warning",
-                    "title": f"{group['label']} signoff attention",
-                    "summary": summary,
-                    "details": attention_details,
-                }
-            )
-        else:
-            status = "ready"
-            summary = (
-                f"{available} of {expected} requirements ready" if expected else "No requirements"
-            )
-        review_groups.append(
-            {
-                "id": group_id,
-                "label": group["label"],
-                "status": status,
-                "available": available,
-                "expected": expected,
-                "summary": summary,
-            }
-        )
-
-    risks.sort(key=lambda risk: risk["severity"] != "blocked")
-    status = checklist_data.get("status")
-    return {
-        "status": status if status in {"ready", "attention", "blocked"} else "blocked",
-        "groups": review_groups,
-        "risks": risks,
-    }
-
-
-def _unavailable_review() -> dict:
-    detail = {
-        "kind": "freshness",
-        "label": "Signoff checklist",
-        "location": "home/checklist.json",
-        "reason": "The current signoff checklist could not be generated.",
-        "owner": "checklist",
-        "policy": "block",
-        "state": "unavailable",
-        "evidence": [],
-    }
-    return {
-        "status": "blocked",
-        "groups": [
-            {
-                "id": group_id,
-                "label": label,
-                "status": "blocked" if group_id == "reports" else "ready",
-                "available": 0,
-                "expected": 0,
-                "summary": "Checklist unavailable" if group_id == "reports" else "No requirements",
-            }
-            for group_id, label in _REVIEW_GROUPS
-        ],
-        "risks": [
-            {
-                "severity": "blocked",
-                "title": "Signoff checklist unavailable",
-                "summary": "Re-run signoff inspection after current-output analysis completes.",
-                "details": [detail],
-            }
-        ],
-    }
-
-
-def _review_detail(item: dict) -> dict:
-    source = item.get("source", {})
-    source = source if isinstance(source, dict) else {}
-    evidence = item.get("evidence", [])
-    return {
-        "kind": item.get("category", "checklist"),
-        "label": item.get("title", "Checklist item"),
-        "location": source.get("path", "home/checklist.json"),
-        "reason": item.get("summary", ""),
-        "owner": item.get("owner", "checklist"),
-        "policy": item.get("policy", "warn"),
-        "state": item.get("state", "unavailable"),
-        "evidence": evidence if isinstance(evidence, list) else [],
-    }
-
-
-def _review_group_for_item(item: dict) -> str:
-    step = str(item.get("step", ""))
-    category = str(item.get("category", ""))
-    source = item.get("source", {})
-    path = source.get("path", "") if isinstance(source, dict) else ""
-    if category == "configuration" or path.startswith("config/"):
-        return "config"
-    if category == "provenance" or path.startswith(("origin/", "initial/")):
-        return "initial"
-    if step == "Harden" or path.startswith(("Harden_ecc/", "harden/")):
-        return "harden"
-    if step == "sta" or path.startswith(("sta_ecc/", "final/timing/sta/")):
-        return "sta"
-    if step == "RCX" or path.startswith(("RCX_ecc/", "final/timing/spef/")):
-        return "spef"
-    if step in {"route", "drc", "lvs", "filler"} or path.startswith(
-        ("route_ecc/", "drc_ecc/", "lvs_ecc/", "filler_ecc/", "final/design/")
-    ):
-        return "final_design"
-    return "reports"
-
-
-def export_signoff_package_archive(
-    workspace,
-    output_path: str,
-    additional_files: list[dict[str, str]] | None = None,
-    *,
-    include_debug: bool = False,
-) -> str:
-    raw_destination = Path(output_path).expanduser()
-    destination = raw_destination.parent.resolve() / raw_destination.name
-
-    with tempfile.TemporaryDirectory(prefix="ecc-signoff-") as temporary_root:
-        result = EngineFlow(workspace).collect_signoff_package(
-            SignoffPackageOptions(
-                output_dir=temporary_root,
-                archive=False,
-                include_debug=include_debug,
-                refresh_analysis=True,
-            )
-        )
-        if not result.ok:
-            missing = ", ".join(result.missing_required) or "unknown required resources"
-            raise RuntimeApiError(
-                "command_failed",
-                f"signoff package is incomplete: {missing}",
-            )
-        if not result.package_dir:
-            raise RuntimeApiError(
-                "command_failed",
-                "signoff package directory was not created",
-            )
-
-        package_dir = Path(result.package_dir)
-
-        if additional_files:
-            package_root = package_dir.resolve()
-            for file_info in additional_files:
-                archive_path = file_info.get("archivePath") if isinstance(file_info, dict) else None
-                content = file_info.get("content") if isinstance(file_info, dict) else None
-                if not isinstance(archive_path, str) or not archive_path:
-                    raise RuntimeApiError(
-                        "invalid_request",
-                        "additionalFiles entries require a non-empty archivePath",
-                    )
-                if not isinstance(content, str):
-                    raise RuntimeApiError(
-                        "invalid_request",
-                        "additionalFiles entries require string content",
-                    )
-                relative_path = Path(archive_path)
-                if (
-                    not relative_path.parts
-                    or relative_path.is_absolute()
-                    or ".." in relative_path.parts
-                ):
-                    raise RuntimeApiError(
-                        "invalid_request",
-                        f"additional file path escapes signoff package: {archive_path}",
-                    )
-                p = package_dir / relative_path
-                try:
-                    p.resolve().relative_to(package_root)
-                except ValueError as exc:
-                    raise RuntimeApiError(
-                        "invalid_request",
-                        f"additional file path escapes signoff package: {archive_path}",
-                    ) from exc
-                p.parent.mkdir(parents=True, exist_ok=True)
-                p.write_text(content, encoding="utf-8")
-
-        archive = package_dir.with_suffix(".tar.gz")
-        import tarfile
-
-        with tarfile.open(archive, "w:gz") as tar:
-            tar.add(package_dir, arcname=package_dir.name)
-
-        destination.parent.mkdir(parents=True, exist_ok=True)
-        descriptor, staged_name = tempfile.mkstemp(
-            dir=destination.parent,
-            prefix=f".{destination.name}.",
-        )
-        os.close(descriptor)
-        staged_path = Path(staged_name)
-        try:
-            shutil.copy2(archive, staged_path)
-            os.replace(staged_path, destination)
-        finally:
-            staged_path.unlink(missing_ok=True)
-
-    return str(destination)
