+++ chipcompiler/tools/ecc/subflow.py
@@ -193,7 +193,7 @@ class EccSubFlow:
 
                 self.save()
 
-                from chipcompiler.runtime.subflow_events import publish_subflow_stage
+                from chipcompiler.engine.subflow_events import publish_subflow_stage
 
                 publish_subflow_stage(self.workspace, self.workspace_step, step_dict)
