+++ chipcompiler/tools/yosys/subflow.py
@@ -109,7 +109,7 @@ class YosysSubFlow:
 
                 self.save()
 
-                from chipcompiler.runtime.subflow_events import publish_subflow_stage
+                from chipcompiler.engine.subflow_events import publish_subflow_stage
 
                 publish_subflow_stage(self.workspace, self.workspace_step, step_dict)
