+++ chipcompiler/tools/yosys_lec/subflow.py
@@ -88,7 +88,7 @@ class YosysLecSubFlow:
                 step_dict["info"] = info
                 self.save()
 
-                from chipcompiler.runtime.subflow_events import publish_subflow_stage
+                from chipcompiler.engine.subflow_events import publish_subflow_stage
 
                 publish_subflow_stage(self.workspace, self.workspace_step, step_dict)
                 break
