+++ pyproject.toml
@@ -23,7 +23,6 @@ dependencies = [
   "ecc-dreamplace==0.1.0a7",
   "ecc-tools-bin==0.1.0a13",
   "fastapi>=0.109",
-  "jsonrpcserver>=5.0.9",
   "klayout>=0.30.2",
   "matplotlib>=3.4",
   "numpy>=1.21",
