--- test/cli/test_rpc_cli.py
@@ -1,32 +1,0 @@
-from chipcompiler.cli import main as cli_main
-
-
-def test_rpc_help_returns_zero_and_lists_serve(capsys):
-    rc = cli_main.run(["rpc", "--help"])
-
-    out = capsys.readouterr().out
-    assert rc == 0
-    assert "serve" in out
-
-
-def test_rpc_serve_help_returns_zero_and_lists_stdio(capsys):
-    rc = cli_main.run(["rpc", "serve", "--help"])
-
-    out = capsys.readouterr().out
-    assert rc == 0
-    assert "--stdio" in out
-    assert "--persistent-db" in out
-
-
-def test_rpc_serve_requires_stdio(capsys):
-    rc = cli_main.run(["rpc", "serve"])
-
-    assert rc != 0
-    assert "--stdio" in capsys.readouterr().err
-
-
-def test_run_help_does_not_list_persistent_db(capsys):
-    rc = cli_main.run(["run", "--help"])
-
-    assert rc == 0
-    assert "--persistent-db" not in capsys.readouterr().out
