+++ test/packaging/test_cli_entrypoint.py
@@ -12,14 +12,14 @@ class TestPackaging:
         assert data["project"]["scripts"]["ecc"] == "chipcompiler.cli.main:main"
         assert set(data["project"]["scripts"]) == {"ecc"}
 
-    def test_pyinstaller_spec_collects_jsonrpcserver_data_files(self):
+    def test_pyinstaller_spec_excludes_removed_jsonrpcserver_data_files(self):
         project_root = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
         spec_path = os.path.join(project_root, "ecc.spec")
 
         with open(spec_path, encoding="utf-8") as f:
             source = f.read()
 
-        assert 'collect_data_files("jsonrpcserver")' in source
+        assert 'collect_data_files("jsonrpcserver")' not in source
 
     def test_pyinstaller_spec_filters_payloads_before_analysis(self):
         project_root = os.path.dirname(os.path.dirname(os.path.dirname(__file__)))
